alx_travel_app/requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==4.2.24
django-cors-headers==4.8.0
django-environ==0.12.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
drf-yasg==1.21.10
idna==3.11
inflection==0.5.1
kombu==5.5.4
mysqlclient==2.2.7
packaging==25.0
prompt_toolkit==3.0.52
pycparser==2.23
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rabbitmq==0.2.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
